// src/types/clinicalHistory.ts
export interface PatientData {
  name: string;
  birthDate: string;
  age: string;
  gender: string;
  reportDate: string;
  folio?: string;
}

export interface VitalSigns {
  weight: string;
  height: string;
  bmi: string;
  bodyFat: string;
  visceralFat: string;
  skeletalMuscle: string;
  temperature: string;
  oxygenSaturation: string;
  heartRate: string;
  respiratoryRate: string;
  bloodPressure: string;
}

export type VisualizationType = 'gauge' | 'bar_chart' | 'simple' | 'list' | 'radar' | 'line_chart' | 'table_row';

export interface RadarPoint {
  name: string;
  value: number;
  unit: string;
  range: string;
}

export interface LabResult {
  name: string;
  value: string | number | string[] | RadarPoint[];
  unit: string;
  range: string;
  description?: string;
  visualizationType: VisualizationType;
  category: string;
  subCategory?: string;
  history?: { date: string; value: number | string }[];
}


export const MEDICAL_SECTIONS = {
  laboratories: 'Laboratorios',
  audiometry: 'Audiometría',
  spirometry: 'Espirometría',
  xRays: 'Rayos X',
  optometry: 'Optometría',
  electrocardiogram: 'Electrocardiograma',
} as const;

export type MedicalSectionId = keyof typeof MEDICAL_SECTIONS;
export type MedicalSectionName = typeof MEDICAL_SECTIONS[MedicalSectionId];

export interface ReportSectionData {
    name: MedicalSectionName;
    summary?: string;
    results: LabResult[];
}

export interface MedicalReport {
  patientData: Partial<PatientData>;
  vitalSigns: Partial<VitalSigns>;
  sections: Partial<Record<MedicalSectionId, ReportSectionData>>;
}